+++ weblisp/src/web.rs
@@ -11,7 +11,6 @@
    hidekuno@gmail.com
 */
 extern crate elisp;
-use crate::buildin;
 use elisp::lisp;
 
 use chrono::Duration;
@@ -23,7 +22,6 @@ use std::fmt;
 use std::fs::File;
 use std::io;
 use std::io::prelude::*;
-use std::path::Path;
 use std::path::PathBuf;
 use std::process::{Child, Command, Stdio};
 
@@ -33,26 +31,24 @@ use log::{debug, error, info, warn};
 pub const PROTOCOL: &str = "HTTP/1.1";
 pub const CRLF: &str = "\r\n";
 
-struct Response(u32, &'static str);
-const RESPONSE_200: Response = Response(200, "OK");
-const RESPONSE_400: Response = Response(400, "Bad Request");
-const RESPONSE_404: Response = Response(404, "Not Found");
-const RESPONSE_405: Response = Response(405, "Method Not Allowed");
-const RESPONSE_500: Response = Response(500, "Internal Server Error");
+pub struct Response(pub u32, pub &'static str);
+pub const RESPONSE_200: Response = Response(200, "OK");
+pub const RESPONSE_400: Response = Response(400, "Bad Request");
+pub const RESPONSE_404: Response = Response(404, "Not Found");
+pub const RESPONSE_405: Response = Response(405, "Method Not Allowed");
+pub const RESPONSE_500: Response = Response(500, "Internal Server Error");
 
-struct MimeType(&'static str, &'static str);
-const MIME_PLAIN: MimeType = MimeType("txt", "text/plain");
-const MIME_HTML: MimeType = MimeType("html", "text/html");
-const MIME_PNG: MimeType = MimeType("png", "image/png");
-const DEFALUT_MIME: &str = "application/octet-stream";
+pub struct MimeType(pub &'static str, pub &'static str);
+pub const MIME_PLAIN: MimeType = MimeType("txt", "text/plain");
+pub const MIME_HTML: MimeType = MimeType("html", "text/html");
+pub const MIME_PNG: MimeType = MimeType("png", "image/png");
+pub const DEFALUT_MIME: &str = "application/octet-stream";
+pub const SESSION_ID: &str = "RUST-ELISP-SID";
+pub const LISP_EXT: &str = ".scm";
+pub type WebResult = (Response, Contents, Option<&'static str>, Option<String>);
+
 const CGI_EXT: &str = ".cgi";
-const LISP_EXT: &str = ".scm";
-
 const LISP: &str = "/lisp";
-const LISP_PARAMNAME: &str = "expr=";
-const SESSION_ID: &str = "RUST-ELISP-SID";
-
-type WebResult = (Response, Contents, Option<&'static str>, Option<String>);
 
 #[derive(Debug)]
 struct UriParseError {
@@ -82,6 +78,7 @@ macro_rules! http_write {
         $s.write_all(CRLF.as_bytes())?;
     };
 }
+#[macro_export]
 macro_rules! http_error {
     ($c: expr) => {
         (
@@ -92,6 +89,7 @@ macro_rules! http_error {
         )
     };
 }
+#[macro_export]
 macro_rules! http_value_error {
     ($c: expr, $err: expr) => {
         (
@@ -108,6 +106,7 @@ macro_rules! http_value_error {
         )
     };
 }
+#[macro_export]
 macro_rules! make_path {
     ($f: expr) => {{
         let mut path = match env::current_dir() {
@@ -129,14 +128,14 @@ pub enum Method {
     Head,
 }
 impl Method {
-    fn as_ref(&self) -> &str {
+    pub fn as_ref(&self) -> &str {
         match self {
             Method::Get => "GET",
             Method::Post => "POST",
             Method::Head => "HEAD",
         }
     }
-    fn create(other: &str) -> Option<Method> {
+    pub fn create(other: &str) -> Option<Method> {
         match other {
             "GET" => Some(Method::Get),
             "POST" => Some(Method::Post),
@@ -209,11 +208,11 @@ impl Request {
         self.get_parsed_list(self.headers.iter().map(|x| x.as_str()).collect(), ':')
     }
 }
-struct WebFile {
+pub struct WebFile {
     file: File,
     length: u64,
 }
-enum Contents {
+pub enum Contents {
     String(String),
     File(WebFile),
     Cgi(Child),
@@ -437,35 +436,16 @@ fn dispatch(r: &Request, env: lisp::Environment, id: usize) -> WebResult {
     return if r.get_resource() == "/" {
         static_contents("index.html")
     } else if r.get_resource().starts_with(LISP) {
-        if !r.get_parameter().starts_with(LISP_PARAMNAME) {
-            return http_error!(RESPONSE_400);
-        }
-        let (_, expr) = r.get_parameter().split_at(LISP_PARAMNAME.len());
-        let mut result = match lisp::do_core_logic(&expr.to_string(), &env) {
-            Ok(r) => r.to_string(),
-            Err(e) => {
-                if lisp::ErrCode::E9000.as_str() == e.get_code() {
-                    env.set_force_stop(false);
-                }
-                e.get_msg()
-            }
-        };
-        result.push_str(CRLF);
-        (
-            RESPONSE_200,
-            Contents::String(result),
-            Some(MIME_PLAIN.1),
-            None,
-        )
+        crate::lisp::do_repl(r, env)
     } else if r.get_resource().ends_with(CGI_EXT) {
         do_cgi(r)
     } else if r.get_resource().ends_with(LISP_EXT) {
-        do_scm(r, env, id)
+        crate::lisp::do_scm(r, env, id)
     } else {
         static_contents(r.get_resource())
     };
 }
-fn set_path_security(path: &mut PathBuf, filename: &str) {
+pub fn set_path_security(path: &mut PathBuf, filename: &str) {
     for s in filename.split('/') {
         if s.is_empty() {
             continue;
@@ -502,7 +482,7 @@ fn static_contents(filename: &str) -> WebResult {
         None,
     )
 }
-fn get_mime(filename: &str) -> &'static str {
+pub fn get_mime(filename: &str) -> &'static str {
     let ext = match filename.rfind('.') {
         Some(i) => &filename[i + 1..],
         None => "",
@@ -533,115 +513,7 @@ fn do_cgi(r: &Request) -> WebResult {
     stdin.write_all(b"\n").unwrap();
     (RESPONSE_200, Contents::Cgi(cgi), None, None)
 }
-fn do_scm(r: &Request, env: lisp::Environment, id: usize) -> WebResult {
-    let path = make_path!(r.get_resource());
-
-    let mut file = match File::open(path) {
-        Ok(file) => file,
-        Err(e) => return http_value_error!(RESPONSE_500, e),
-    };
-    let path = Path::new(r.get_resource());
-    let f = match path.file_name() {
-        Some(f) => match f.to_str() {
-            Some(f) => f.replace(LISP_EXT, ""),
-            None => return http_error!(RESPONSE_500),
-        },
-        None => return http_error!(RESPONSE_500),
-    };
-
-    let mut load_file = String::new();
-    if file.read_to_string(&mut load_file).is_err() {
-        return http_error!(RESPONSE_500);
-    }
-    match lisp::do_core_logic(&load_file, &env) {
-        Ok(_) => {}
-        Err(e) => {
-            error!("{}", e.get_msg());
-            return http_value_error!(RESPONSE_500, e.get_msg());
-        }
-    };
-
-    let method = if let Some(method) = r.get_method() {
-        method.as_ref().to_string()
-    } else {
-        String::from("")
-    };
-    let (sid, first) = get_session_id(r, id);
-
-    let lisp = format!(
-        "((lambda () ({}::main #({:#?} {} {} {:#?} {:#?}) {:#?})))",
-        f,
-        method,
-        r.get_lisp_header(),
-        r.get_lisp_param(),
-        r.get_resource(),
-        r.get_protocol(),
-        sid,
-    );
-
-    let result = match lisp::do_core_logic(&lisp, &env) {
-        Ok(v) => v,
-        Err(e) => {
-            error!("{}", e.get_msg());
-            return http_value_error!(RESPONSE_500, e.get_msg());
-        }
-    };
-    let cookie = if first { Some(sid) } else { None };
-    parse_lisp_result(result, env, cookie)
-}
-fn get_session_id(r: &Request, id: usize) -> (String, bool) {
-    for e in r.get_headers() {
-        if e.starts_with("Cookie:") && e.contains(SESSION_ID) {
-            // Cookie: RUST-ELISP-SID=RE-1641713444-1
-            let s = "Cookie: RUST-ELISP-SID=";
-            if e.contains(s) {
-                return (e[s.len()..].to_string(), false);
-            }
-        }
-    }
-    let session_id = format!("RE-{}-{}", Utc::now().timestamp(), id);
-    (session_id, true)
-}
-fn parse_lisp_result(
-    exp: lisp::Expression,
-    env: lisp::Environment,
-    cookie: Option<String>,
-) -> WebResult {
-    let l = match lisp::eval(&exp, &env) {
-        Ok(v) => match v {
-            lisp::Expression::Vector(l) => l,
-            _ => return http_error!(RESPONSE_500),
-        },
-        Err(e) => return http_value_error!(RESPONSE_500, e.get_msg()),
-    };
-    let l = &*(elisp::referlence_list!(l));
-    if l.len() != buildin::RESPONSE_COLUMNS {
-        return http_error!(RESPONSE_500);
-    }
-
-    let status = match lisp::eval(&l[0], &env) {
-        Ok(v) => match v {
-            lisp::Expression::Integer(i) => i,
-            _ => return http_error!(RESPONSE_500),
-        },
-        Err(e) => return http_value_error!(RESPONSE_500, e.get_msg()),
-    };
-
-    let mime = match lisp::eval(&l[1], &env) {
-        Ok(v) => match v {
-            lisp::Expression::String(s) => s,
-            _ => return http_error!(RESPONSE_500),
-        },
-        Err(e) => return http_value_error!(RESPONSE_500, e.get_msg()),
-    };
-    (
-        get_status(status),
-        Contents::String(l[2].to_string()),
-        Some(get_mime(&(".".to_owned() + &mime))),
-        cookie,
-    )
-}
-fn get_status(status: i64) -> Response {
+pub fn get_status(status: i64) -> Response {
     match status as u32 {
         200 => RESPONSE_200,
         400 => RESPONSE_400,
